+++ Feelter/Sources/Feature/SignUp/SignUpViewModel.swift
@@ -133,16 +133,3 @@ final class SignUpViewModel: ViewModel {
         return output
     }
 }
-
-// MARK: - Model
-extension SignUpViewModel {
-    struct SignUpForm {
-        var email: String = ""
-        var password: String = ""
-        var nickname: String = ""
-        var name: String?
-        var phoneNumber: String?
-        var introduction: String?
-        var hashTags: [String] = []
-    }
-}
